feat: Product Detail 조회시 즐겨찾기 수 조회 및 즐겨찾기 삭제 구현

+++ backend/Business/src/main/java/com/picky/business/product/dto/ProductDetailResponse.java
@@ -15,7 +15,7 @@ public class ProductDetailResponse {
     private String filename;
     private String badge;
     private int category;
-    private int favoriteCount; //좋아요 수
+    private long favoriteCount; //좋아요 수
     private int weight;
     private int kcal;
     private double carb;
